Fix style toggle with deckGL layer
